refactor: break singluar alpaca_rest provider code file into smaller modules
