+++ src/market_data_ingestor/src/models/request_params.rs
@@ -1,7 +1,7 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
-use crate::{models::{asset::AssetClass, timeframe::TimeFrame}, providers::alpaca_rest::AlpacaBarsParams};
+use crate::{models::{asset::AssetClass, timeframe::TimeFrame}, providers::alpaca_rest::params::AlpacaBarsParams};
 
 /// Universal parameters for requesting time-series bar data from any market data provider.
 ///
