+++ src/market_data_ingestor/src/providers/alpaca_rest/provider.rs
@@ -1,85 +1,13 @@
 use async_trait::async_trait;
-use chrono::{DateTime, Utc};
-use indexmap::IndexMap;
 use reqwest::{header, Client};
 use secrecy::{ExposeSecret, SecretString};
-use serde::{Deserialize, Serialize};
 use shared_utils::env::get_env_var;
 
-use crate::{models::{bar::Bar, bar_series::BarSeries, request_params::{BarsRequestParams, ProviderParams}, timeframe::{TimeFrame, TimeFrameUnit}}, providers::{DataProvider, ProviderError, ProviderInitError}};
+use crate::{models::{bar::Bar, bar_series::BarSeries, request_params::{BarsRequestParams, ProviderParams}, timeframe::{TimeFrame, TimeFrameUnit}}, providers::{alpaca_rest::{params::AlpacaBarsParams, response::AlpacaResponse}, DataProvider, ProviderError, ProviderInitError}};
+
+
 
 const BASE_URL: &str = "https://data.alpaca.markets/v2/stocks/bars";
-
-/// Specifies the corporate action adjustment for stock data.
-#[derive(Clone, Debug, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum Adjustment {
-    #[default]
-    Raw,
-    Split,
-    Dividend,
-    All,
-}
-
-/// Specifies the source feed for stock data.
-#[derive(Clone, Debug, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum Feed {
-    #[default]
-    Sip,
-    Iex,
-    Otc,
-}
-
-/// Specifies the sort order for the bars.
-#[derive(Clone, Debug, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum Sort {
-    #[default]
-    Asc,
-    Desc,
-}
-
-/// Alpaca-specific parameters for a bars request.
-#[derive(Clone, Debug, Serialize, Deserialize, Default)]
-pub struct AlpacaBarsParams {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub adjustment: Option<Adjustment>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub feed: Option<Feed>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub currency: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sort: Option<Sort>,
-}
-
-#[derive(Deserialize, Debug)]
-struct AlpacaBar {
-    #[serde(rename = "t")]
-    timestamp: DateTime<Utc>,
-    #[serde(rename = "o")]
-    open: f64,
-    #[serde(rename = "h")]
-    high: f64,
-    #[serde(rename = "l")]
-    low: f64,
-    #[serde(rename = "c")]
-    close: f64,
-    #[serde(rename = "v")]
-    volume: f64,
-    #[serde(rename = "n")]
-    trade_count: u64,
-    #[serde(rename = "vw")]
-    vwap: f64,
-}
-
-#[derive(Deserialize, Debug)]
-struct AlpacaResponse {
-    bars: IndexMap<String, Vec<AlpacaBar>>,
-    next_page_token: Option<String>,
-}
 
 pub struct AlpacaProvider {
     client: Client,
